# docs/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: team@testseer.io
    name: TestSeer Team
  description: "TestSeer analyses Java service codebases to produce a structured knowledge\
    \ graph of classes, endpoints, and outbound HTTP calls. Use the Query API to ask\
    \ questions about your services — reachability, impact analysis, shared type consumers\
    \ — and always get freshness metadata alongside results."
  license:
    name: Proprietary
  title: TestSeer API
  version: 0.1.0
servers:
- description: Local development
  url: http://localhost:8080
...
paths:
...
  /v1/facts/data-access:
    get:
      operationId: getDataAccess
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListDataAccessView"
          description: OK
      summary: Data access facts along event handlers (C-P3)
      tags:
      - Query — Messaging Flow
...
  /v1/facts/gates:
    get:
      operationId: getGates
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: env
        required: false
        schema:
          type: string
      - in: query
        name: flowStep
        required: false
        schema:
          type: string
      - in: query
        name: refreshLive
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListFlowGateView"
          description: OK
      summary: Flow gate and config facts (C-P6)
      tags:
      - Query — Messaging Flow
...
  /v1/facts/message-schema:
    get:
      operationId: getMessageSchemas
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: topicShortId
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListMessageSchemaView"
          description: OK
      summary: Message schema facts (C-P2)
      tags:
      - Query — Messaging Flow
...
  /v1/facts/pubsub:
    get:
      operationId: getPubSubInventory
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: env
        required: false
        schema:
          type: string
      - in: query
        name: shortId
        required: false
        schema:
          type: string
      - in: query
        name: role
        required: false
        schema:
          type: string
      - in: query
        name: liveVerify
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListPubSubView"
          description: OK
      summary: Pub/Sub resource inventory (C-P1)
      tags:
      - Query — Messaging Flow
  /v1/facts/pubsub/org:
    get:
      operationId: getOrgPubSubInventory
      parameters:
      - in: query
        name: orgId
        required: true
        schema:
          type: string
      - in: query
        name: env
        required: false
        schema:
          type: string
      - in: query
        name: resourceKind
        required: false
        schema:
          type: string
      - in: query
        name: role
        required: false
        schema:
          type: string
      - in: query
        name: liveVerify
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListPubSubOrgView"
          description: OK
      summary: Org-wide Pub/Sub inventory for cross-repo flows
      tags:
      - Query — Messaging Flow
  /v1/facts/validation-hints:
    get:
      operationId: getValidationHints
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: env
        required: false
        schema:
          type: string
      - in: query
        name: flowStep
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListValidationHintView"
          description: OK
      summary: Validation hints for test plan Expected columns (C-P4)
      tags:
      - Query — Messaging Flow
...
  /v1/gaps/messaging:
    get:
      operationId: getMessagingGaps
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: env
        required: false
        schema:
          type: string
          default: pdn
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeEventFlowReport"
          description: OK
      summary: Messaging flow gap report (C-P5)
      tags:
      - Query — Messaging Flow
...
  /v1/graph/event-flow:
    get:
      operationId: traceEventFlow
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: shortId
        required: false
        schema:
          type: string
      - in: query
        name: env
        required: false
        schema:
          type: string
          default: pdn
      - in: query
        name: includeExternal
        required: false
        schema:
          type: boolean
          default: false
      - in: query
        name: liveVerify
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeEventFlowReport"
          description: OK
      summary: "Trace topic through publishers, schemas, DB, gates (C-P2/P4/P6)"
      tags:
      - Query — Messaging Flow
  /v1/graph/event-flow/cross-repo:
    get:
      operationId: traceCrossRepoFlow
      parameters:
      - in: query
        name: orgId
        required: true
        schema:
          type: string
      - in: query
        name: shortId
        required: true
        schema:
          type: string
      - in: query
        name: env
        required: false
        schema:
          type: string
          default: pdn
      - in: query
        name: maxHops
        required: false
        schema:
          type: integer
          format: int32
          default: 12
      - in: query
        name: bundle
        required: false
        schema:
          type: string
      - in: query
        name: includeExternal
        required: false
        schema:
          type: boolean
          default: false
      - in: query
        name: liveVerify
        required: false
        schema:
          type: boolean
          default: false
      - in: query
        name: followMode
        required: false
        schema:
          type: string
          default: runtime
      - in: query
        name: includeManifest
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeCrossRepoFlowReport"
          description: OK
      summary: Trace topic across all indexed services in an org (cross-repo)
      tags:
      - Query — Messaging Flow
...
  /v1/jobs/{jobId}:
    get:
      description: Returns analysis_runs lifecycle fields for observability and debugging.
      operationId: getJob
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
          description: Job found
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
          description: Job not found
      summary: Get ingestion job status by jobId
      tags:
      - Query — Jobs
...
  /v1/status/{serviceId}:
    get:
      description: "Returns the freshness status (CURRENT, STALE, INDEXING, NOT_INDEXED)\
        \ for the given service, along with the last indexed commit SHA and timestamp.\
        \ This endpoint is never cached — it always reads live from analysis_runs."
      operationId: status
      parameters:
      - description: Unique service identifier
        in: path
        name: serviceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeMapStringObject"
          description: Status returned (check freshnessStatus field)
      summary: Get indexing status for a service
      tags:
      - Query — Status
...
components:
  schemas:
...
    ConsistencyHintView:
      type: object
      properties:
        confidence:
          type: number
          format: double
        correlationKeys:
          type: array
          items:
            type: string
        downstreamGates:
          type: array
          items:
            $ref: "#/components/schemas/DownstreamGateView"
        evidenceSource:
          type: string
        invariants:
          type: array
          items:
            $ref: "#/components/schemas/ConsistencyInvariantHintView"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ConsistencyParticipantHintView"
        pattern:
          type: string
        pollStrategy:
          $ref: "#/components/schemas/ConsistencyPollStrategyView"
        primaryPhysical:
          type: string
        primaryStore:
          type: string
        scenarioId:
          type: string
    ConsistencyInvariantHintView:
      type: object
      properties:
        description:
          type: string
        kind:
          type: string
        pollHint:
          type: string
...
    DataAccessView:
      type: object
      properties:
        accessorFqn:
          type: string
        accessorKind:
          type: string
        catalogRef:
          type: string
        confidence:
          type: number
          format: double
        consistencyHints:
          type: array
          items:
            $ref: "#/components/schemas/ConsistencyHintView"
        correlationKeys:
          type: string
        daoMethod:
          type: string
        domainFqn:
          type: string
        entityFqn:
          type: string
        evidenceSource:
          type: string
        flowSteps:
          type: string
        handlerClassFqn:
          type: string
        handlerMethod:
          type: string
        operation:
          type: string
        physicalName:
          type: string
        pollHint:
          type: string
        secondaryStores:
          type: string
        storeType:
          type: string
        tableOrEntity:
          type: string
        validationKind:
          type: string
...
    EventFlowReport:
      type: object
      properties:
        envLane:
          type: string
        gaps:
          type: array
          items:
            $ref: "#/components/schemas/FlowGap"
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        pubsubResources:
          type: array
          items:
            $ref: "#/components/schemas/PubSubView"
        serviceId:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/EventFlowStep"
        topicShortId:
          type: string
...
    FlowGateView:
      type: object
      properties:
        confidence:
          type: number
          format: double
        effectWhenFail:
          type: string
        evidenceSource:
          type: string
        gateKey:
          type: string
        gateKind:
          type: string
        guardedFlowStep:
          type: string
        guardedSymbolFqn:
          type: string
        liveEvidence:
          type: string
        liveSnapshotAt:
          type: string
          format: date-time
        liveStatus:
          type: string
        liveValue:
          type: string
        requiredValue:
          type: string
        testPrecondition:
          type: string
...
    JobStatusResponse:
      type: object
      properties:
        attempt:
          type: integer
          format: int32
        commitSha:
          type: string
        completedAt:
          type: string
          format: date-time
        enqueuedAt:
          type: string
          format: date-time
        errorDetail:
          type: string
        jobId:
          type: string
        jobType:
          type: string
        orgId:
          type: string
        serviceId:
          type: string
        startedAt:
          type: string
          format: date-time
        status:
          type: string
...
    MessageSchemaView:
      type: object
      properties:
        confidence:
          type: number
          format: double
        direction:
          type: string
        evidenceSource:
          type: string
        linkedClassFqn:
          type: string
        payloadFields:
          type: string
        payloadProto:
          type: string
        topicShortId:
          type: string
        unpackExpression:
          type: string
...
    PubSubOrgView:
      type: object
      properties:
        confidence:
          type: number
          format: double
        consistencyHints:
          type: array
          items:
            $ref: "#/components/schemas/ConsistencyHintView"
        envLane:
          type: string
        evidenceSource:
          type: string
        fullResourceId:
          type: string
        inboundTriggers:
          type: array
          items:
            $ref: "#/components/schemas/EntryTriggerView"
        linkedClassFqn:
          type: string
        liveVerification:
          $ref: "#/components/schemas/LiveVerificationView"
        moduleName:
          type: string
        repo:
          type: string
        resourceKind:
          type: string
        role:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        shortId:
          type: string
        springKey:
          type: string
        transport:
          type: string
        workloadName:
          type: string
    PubSubView:
      type: object
      properties:
        confidence:
          type: number
          format: double
        envLane:
          type: string
        evidenceSource:
          type: string
        fullResourceId:
          type: string
        linkedClassFqn:
          type: string
        liveVerification:
          $ref: "#/components/schemas/LiveVerificationView"
        moduleName:
          type: string
        resourceKind:
          type: string
        role:
          type: string
        shortId:
          type: string
        springKey:
          type: string
        transport:
          type: string
        workloadName:
          type: string
...
    ResponseEnvelopeCrossRepoFlowReport:
      type: object
      properties:
        commitSha:
          type: string
        data:
          $ref: "#/components/schemas/CrossRepoFlowReport"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopeEventFlowReport:
      type: object
      properties:
        commitSha:
          type: string
        data:
          $ref: "#/components/schemas/EventFlowReport"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopeListDataAccessView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/DataAccessView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopeListFlowGateView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/FlowGateView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopeListMessageSchemaView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/MessageSchemaView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopeListPubSubOrgView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/PubSubOrgView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
    ResponseEnvelopeListPubSubView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/PubSubView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopeListValidationHintView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ValidationHintView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
    ResponseEnvelopeMapStringObject:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: object
          additionalProperties:
            type: object
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ValidationHintView:
      type: object
      properties:
        envLane:
          type: string
        flowStep:
          type: string
        hintKind:
          type: string
        hintValue:
          type: string
        linkedSymbolFqn:
          type: string
...